LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/generic-family-keywords-003-expected.txt


PASS @font-face matching for quoted and unquoted serif (drawing text in a canvas)
PASS @font-face matching for quoted and unquoted sans-serif (drawing text in a canvas)
PASS @font-face matching for quoted and unquoted cursive (drawing text in a canvas)
PASS @font-face matching for quoted and unquoted fantasy (drawing text in a canvas)
PASS @font-face matching for quoted and unquoted monospace (drawing text in a canvas)
FAIL @font-face matching for quoted and unquoted system-ui (drawing text in a canvas) assert_equals: quoted system-ui matches  @font-face rule expected 125 but got 28.381288528442383
FAIL @font-face matching for quoted and unquoted math (drawing text in a canvas) assert_equals: quoted math matches  @font-face rule expected 125 but got 33.375
FAIL @font-face matching for quoted and unquoted generic(fangsong) (drawing text in a canvas) promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL @font-face matching for quoted and unquoted generic(kai) (drawing text in a canvas) promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL @font-face matching for quoted and unquoted generic(khmer-mul) (drawing text in a canvas) promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL @font-face matching for quoted and unquoted generic(nastaliq) (drawing text in a canvas) promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL @font-face matching for quoted and unquoted ui-serif (drawing text in a canvas) assert_equals: quoted ui-serif matches  @font-face rule expected 125 but got 38.3699836730957
FAIL @font-face matching for quoted and unquoted ui-sans-serif (drawing text in a canvas) assert_equals: quoted ui-sans-serif matches  @font-face rule expected 125 but got 28.381288528442383
FAIL @font-face matching for quoted and unquoted ui-monospace (drawing text in a canvas) assert_equals: quoted ui-monospace matches  @font-face rule expected 125 but got 77.2705078125
FAIL @font-face matching for quoted and unquoted ui-rounded (drawing text in a canvas) assert_equals: quoted ui-rounded matches  @font-face rule expected 125 but got 42.05322265625
